refactor sources and country in feed settings
