+++ application/lib/presentation/navigation/pages.dart
@@ -7,9 +7,9 @@ import 'package:xayn_discovery_app/presentation/bookmark/widget/bookmarks_screen
 import 'package:xayn_discovery_app/presentation/discovery_card/screen/discovery_card_screen.dart';
 import 'package:xayn_discovery_app/presentation/discovery_feed/widget/discovery_feed.dart';
 import 'package:xayn_discovery_app/presentation/error/widget/error_screen.dart';
-import 'package:xayn_discovery_app/presentation/feed_settings/page/country_feed_settings_page.dart';
-import 'package:xayn_discovery_app/presentation/feed_settings/page/source/widget/add_source_screen.dart';
-import 'package:xayn_discovery_app/presentation/feed_settings/page/source/widget/sources_screen.dart';
+import 'package:xayn_discovery_app/presentation/feed_settings/country/widget/country_feed_settings_page.dart';
+import 'package:xayn_discovery_app/presentation/feed_settings/source/widget/add_source_screen.dart';
+import 'package:xayn_discovery_app/presentation/feed_settings/source/widget/sources_screen.dart';
 import 'package:xayn_discovery_app/presentation/payment/paywall_screen.dart';
 import 'package:xayn_discovery_app/presentation/personal_area/personal_area_screen.dart';
 import 'package:xayn_discovery_app/presentation/settings/settings_screen.dart';
