+++ application/test/presentation/feed_settings/page/select_countries_test.dart
@@ -3,8 +3,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:xayn_design/src/utils/design_testing_utils.dart';
 import 'package:xayn_discovery_app/domain/model/country/country.dart';
 import 'package:xayn_discovery_app/presentation/constants/r.dart';
-import 'package:xayn_discovery_app/presentation/feed_settings/page/country_feed_settings_page.dart';
-import 'package:xayn_discovery_app/presentation/feed_settings/widget/country_item.dart';
+import 'package:xayn_discovery_app/presentation/feed_settings/country/widget/country_feed_settings_page.dart';
+import 'package:xayn_discovery_app/presentation/feed_settings/country/widget/country_item.dart';
 
 import '../../../test_utils/extensions.dart';
